// src/audit/profiles/audit-body-profile.enum.ts
export enum AuditEvent {
  USER_UPDATE = 'USER_UPDATE',
  USER_DELETE = 'USER_DELETE',

  RESERVATION_CREATE = 'RESERVATION_CREATE',
  RESERVATION_DELETE = 'RESERVATION_DELETE',

  AUTH_RESERVATION_VERIFICATION = 'AUTH_RESERVATION_VERIFICATION',
  AUTH_RESET_PASSWORD = 'AUTH_RESET_PASSWORD',
  AUTH_LOGIN_TO_SYSTEM = 'AUTH_LOGIN_TO_SYSTEM',

  APPO_PATCH = 'APPO_PATCH',
  APPO_DELETE = 'APPO_DELETE',
  APPO_BY_DAY = 'APPO_BY_DAY',
  APPO_BY_HOUR = 'APPO_BY_HOUR',
  APPO_ADM_MODIFIES_USERS = 'APPO_ADM_MODIFIES_USERS',
  APPO_INS_MODIFIES = 'APPO_INS_MODIFIES',

  REMOVE_EXPIRED_TOKENS = 'REMOVE_EXPIRED_TOKENS',
}